// frontend/src/api/generated/model/jobStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * print-web
 * OpenAPI spec version: 0.1.0
 */

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  PRINTING: 'PRINTING',
  DONE: 'DONE',
  FAILED: 'FAILED',
  REJECTED: 'REJECTED',
} as const;